Made changes in main activity (room, mvvm and retrofit status OK)
