+++ app/src/main/java/com/example/cinemax/localdata/LocalViewModel.kt
@@ -1,0 +1,9 @@
+package com.example.cinemax.localdata
+
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+
+class LocalViewModel {
+
+}
