+++ app/src/main/java/com/example/cinemax/models/MovieResponse.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 
 data class MovieResponse (
     @SerializedName("page") val page: Int,
-    @SerializedName("results") val movies: List<MovieEntity>,
+    @SerializedName("results") val movies: List<Movie>,
     @SerializedName("total_pages") val pages: Int
 )
